+++ Controls/_editInPlace/Controller.ts
@@ -1,9 +1,9 @@
 import {Model, DestroyableMixin} from 'Types/entity';
 import {Logger} from 'UI/Utils';
-import {IEditInPlace, IEditInPlaceOptions, TAsyncOperationResult} from './interface/IEditInPlace';
 import {CONSTANTS, TEditableCollectionItem} from './Types';
 import {CollectionEditor} from './CollectionEditor';
 import {mixin} from 'Types/util';
+import {Collection} from 'Controls/display';
 
 const ERROR_MSG = {
     COLLECTION_IS_NOT_DEFINED: 'IEditInPlaceOptions.collection is not defined. Option is required. It must be installed at least once.',
@@ -13,9 +13,88 @@ const ERROR_MSG = {
 };
 
 /**
+ * @typedef {Promise.<void | { canceled: true }>} TAsyncOperationResult
+ * @description Тип возвращаемого значения из операций редактирования по месту.
+ */
+type TAsyncOperationResult = Promise<void | { canceled: true }>;
+
+/**
+ * @typedef {void|CONSTANTS.CANCEL|Promise.<void|{CONSTANTS.CANCEL}>} TBeforeCallbackBaseResult
+ * @description Базовый тип, который можно вернуть из любой функций обратного вызова до начала операции редактирования по месту.
+ */
+type TBeforeCallbackBaseResult = void | CONSTANTS.CANCEL | Promise<void | CONSTANTS.CANCEL>;
+
+/**
+ * @typedef {Function} TBeforeBeginEditCallback
+ * @description Функция обратного вызова перед запуском редактирования.
+ * @param {Object} options Набор опций для запуска редактирования. Доступные свойства: item {Types/entity:Model} - запись для которой запускается редактирование.
+ * @param {Boolean} isAdd Флаг, принимает значение true, если запись добавляется
+ */
+type TBeforeBeginEditCallback = (options: { item?: Model }, isAdd: boolean) =>
+    TBeforeCallbackBaseResult | { item?: Model } | Promise<{ item?: Model }>;
+
+/**
+ * @typedef {Function} TBeforeEndEditCallback
+ * @description Функция обратного вызова перед завершением редактирования
+ * @param {Types/entity:Model} item Редактируемая запись для которой запускается завершение редактирования.
+ * @param willSave Флаг, принимает значение true, если ожидается, что запись будет сохранена.
+ * @param isAdd Флаг, принимает значение true, если запись добавляется
+ */
+type TBeforeEndEditCallback = (item: Model, willSave: boolean, isAdd: boolean) => TBeforeCallbackBaseResult;
+
+/**
+ * Интерфейс опций контроллера редактирования по месту.
+ * @interface Controls/_editInPlace/IEditInPlaceOptions
+ * @public
+ * @author Родионов Е.А.
+ */
+/*
+ * Interface of edit in place controller options.
+ * @interface Controls/_editInPlace/IEditInPlaceOptions
+ * @public
+ * @author Rodionov E/
+ */
+interface IEditInPlaceOptions {
+    /**
+     * @name Controls/_editInPlace/IEditInPlaceOptions#collection
+     * @cfg {Collection.<Types/entity:Model>} Коллекция элементов.
+     */
+    collection: Collection<Model>;
+
+    /**
+     * @name Controls/_editInPlace/IEditInPlaceOptions#onBeforeBeginEdit
+     * @cfg {TBeforeBeginEditCallback} Функция обратного вызова перед запуском редактирования.
+     */
+    onBeforeBeginEdit?: TBeforeBeginEditCallback;
+
+    /**
+     * @name Controls/_editInPlace/IEditInPlaceOptions#onAfterBeginEdit
+     * @cfg {Function} Функция обратного вызова после запуска редактирования.
+     * @param {Types/entity:Model} item Запись для которой запускается редактирование.
+     * @param {Boolean} isAdd Флаг, принимает значение true, если запись добавляется.
+     * @void
+     */
+    onAfterBeginEdit?: (item: Model, isAdd: boolean) => void;
+
+    /**
+     * @name Controls/_editInPlace/IEditInPlaceOptions#onBeforeEndEdit
+     * @cfg {TBeforeEndEditCallback} Функция обратного вызова перед завершением редактирования.
+     */
+    onBeforeEndEdit?: TBeforeEndEditCallback;
+
+    /**
+     * @name Controls/_editInPlace/IEditInPlaceOptions#onAfterEndEdit
+     * @cfg {Function} Функция обратного вызова после завершением редактирования.
+     * @param {Types/entity:Model} item Редактируемая запись.
+     * @param {Boolean} isAdd Флаг, принимает значение true, если запись добавлялась.
+     * @void
+     */
+    onAfterEndEdit?: (item: Model, isAdd: boolean) => void;
+}
+
+/**
  * Контроллер редактирования по месту.
  *
- * @mixes Controls/_editInPlace/interface/IEditInPlace
  * @mixes Types/_entity/DestroyableMixin
  * @public
  * @class Controls/_editInPlace/Controller
@@ -25,13 +104,12 @@ const ERROR_MSG = {
 /*
  * Edit in place controller.
  *
- * @mixes Controls/_editInPlace/interface/IEditInPlace
  * @mixes Types/_entity/DestroyableMixin
  * @public
  * @class Controls/_editInPlace/Controller
- * @author Rodionov E.
+ * @author Rodionov E.A.
  */
-export class Controller extends mixin<DestroyableMixin>(DestroyableMixin) implements IEditInPlace {
+export class Controller extends mixin<DestroyableMixin>(DestroyableMixin) {
     private _options: IEditInPlaceOptions;
     private _collectionEditor: CollectionEditor;
     private _operationsPromises: {
@@ -44,6 +122,15 @@ export class Controller extends mixin<DestroyableMixin>(DestroyableMixin) implem
         this.updateOptions(options);
     }
 
+    /**
+     * Обновить опции контроллера.
+     * @method
+     * @param {Partial.<Controls/_editInPlace/IEditInPlaceOptions>} newOptions Новые опции.
+     * @void
+     *
+     * @public
+     * @remark Все поля в новых опциях не являются обязательными, таким образом, есть возможность выборочного обновления.
+     */
     updateOptions(newOptions: Partial<IEditInPlaceOptions>): void {
         const oldOptions: Partial<IEditInPlaceOptions> = this._options || {};
         const hasCollection = !!newOptions.collection || !!oldOptions.collection;
@@ -69,30 +156,100 @@ export class Controller extends mixin<DestroyableMixin>(DestroyableMixin) implem
         }
     }
 
+    /**
+     * Получить редактируемый элемент
+     * @method
+     * @return {Types/entity:Model|undefined}
+     * @public
+     */
     getEditingItem(): Model | undefined {
         return this._collectionEditor.getEditingItem();
     }
 
+    /**
+     * Начинать добавление переданного элемента. Если элемент не передан, ожидается что он будет возвращен из функции обратного вызова IEditInPlaceOptions.onBeforeBeginEdit.
+     * @method
+     * @param {Types/entity:Model|undefined} item Элемент для добавления
+     * @param {TAddPosition} addPosition позиция добавляемого элемента
+     * @return {TAsyncOperationResult}
+     *
+     * @public
+     * @async
+     *
+     * @see Controls/_editInPlace/IEditInPlaceOptions#onBeforeBeginEdit
+     * @see Controls/_editInPlace/IEditInPlaceOptions#onAfterBeginEdit
+     *
+     * @remark Запуск добавления может быть отменен. Для этого из функции обратного вызова IEditInPlaceOptions.onBeforeBeginEdit необхобимо вернуть константу отмены.
+     */
     add(item?: Model | undefined, addPosition: 'top' | 'bottom' = 'bottom'): TAsyncOperationResult {
         return this._endPreviousAndBeginEdit(item, true, addPosition);
     }
 
+    /**
+     * Запустить редактирование переданного элемента. Если элемент не передан, ожидается что он будет возвращен из функции обратного вызова IEditInPlaceOptions.onBeforeBeginEdit.
+     * @method
+     * @param {Types/entity:Model|undefined} item Элемент для редактирования
+     * @return {TAsyncOperationResult}
+     *
+     * @public
+     * @async
+     *
+     * @remark Запуск редактирования может быть отменен. Для этого из функции обратного вызова IEditInPlaceOptions.onBeforeBeginEdit необхобимо вернуть константу отмены.
+     */
     edit(item?: Model): TAsyncOperationResult {
         return this._endPreviousAndBeginEdit(item, false);
     }
 
+    /**
+     * Завершить редактирование элемента и сохранить изменения.
+     * @method
+     * @return {TAsyncOperationResult}
+     *
+     * @public
+     * @async
+     *
+     * @see Controls/_editInPlace/IEditInPlaceOptions#onBeforeEndEdit
+     * @see Controls/_editInPlace/IEditInPlaceOptions#onAfterEndEdit
+     *
+     * @remark Завершение редактирования может быть отменено. Для этого из функции обратного вызова IEditInPlaceOptions.onBeforeEndEdit необхобимо вернуть константу отмены.
+     */
     commit(): TAsyncOperationResult {
         return this._endEdit(true);
     }
 
+    /**
+     * Завершить редактирование элемента и отменить изменения.
+     * @method
+     * @return {TAsyncOperationResult}
+     *
+     * @public
+     * @async
+     *
+     * @see Controls/_editInPlace/IEditInPlaceOptions#onBeforeEndEdit
+     * @see Controls/_editInPlace/IEditInPlaceOptions#onAfterEndEdit
+     *
+     * @remark Завершение редактирования может быть отменено. Для этого из функции обратного вызова IEditInPlaceOptions.onBeforeEndEdit необхобимо вернуть константу отмены.
+     */
     cancel(): TAsyncOperationResult {
         return this._endEdit(false);
     }
 
+    /**
+     * Получить следующий элемент коллекции, для которого доступно редактирование.
+     * @method
+     * @return {CollectionItem.<Types/entity:Model>|undefined}
+     * @public
+     */
     getNextEditableItem(): TEditableCollectionItem {
         return this._collectionEditor.getNextEditableItem();
     }
 
+    /**
+     * Получить предыдущий элемент коллекции, для которого доступно редактирование.
+     * @method
+     * @return {CollectionItem.<Types/entity:Model>|undefined}
+     * @public
+     */
     getPrevEditableItem(): TEditableCollectionItem {
         return this._collectionEditor.getPrevEditableItem();
     }
