+++ Controls/editInPlace.ts
@@ -11,6 +11,7 @@
  * Library that provides edit in place for collection
  * @library Controls/editInPlace
  * @includes Controller Controls/_editInPlace/Controller
+ *
  * @public
  * @author Родионов Е.А.
  */
